добавляет новый модуль с валидацией
