+++ js/main.js
@@ -13,8 +13,6 @@ openForm();
 import { closeForm } from './form.js';
 closeForm();
 import './form.js';
-import { onload } from './form.js';
+import { onload } from './validation.js';
 onload();
-import { hashtagCheck } from './regexp.js';
-hashtagCheck();
