+++ js/regexp.js
@@ -1,12 +1,10 @@
 const textHashtags = document.querySelector('.text__hashtags');
-const uploadForm = document.querySelector('.img-upload__form');
+
 
 const hashtagCheck = function () {
-  const pristine = new Pristine(uploadForm);
 
   const regexp = /^(#[a-zа-яё0-9]{1,19}\s){0,4}(#[a-zа-яё0-9]{1,19})(?![\s\S]*#\1)$/i;
   regexp.test(textHashtags.value);
-  pristine.validate();
   if (regexp.test(textHashtags.value)) {
     const hashtagArray = Array.from(textHashtags.value);
 
@@ -19,16 +17,7 @@ const hashtagCheck = function () {
     }
   }
 
-
-  uploadForm.addEventListener('submit', (evt) => {
-    evt.preventDefault();
-
-    if (hashtagCheck) {
-      uploadForm.submit();
-    }
-  });
 };
-
 hashtagCheck();
 export {hashtagCheck};
